requirements.txt
torch==2.3.0
numpy==1.26.4
azureml-core==1.56.0
azureml-inference-server-http==1.4.0
azure-functions==1.21.1
requests==2.32.3
azure-iot-device==2.13.2
azure-cosmos==4.7.0